KAMP univariate variance

Computes the KAMP (K-function Adjusted for Marked Permutations) variance for a given spatial point pattern. Also returns the KAMP expectation, z-statistic, and p-value.

Note: this a matrix-based implementation of the KAMP variance that does not use the spatstat package. It is a wrapper around the kamp_variance_helper function that calculates the KAMP variance at one radius and maps it over a vector of radii.

Usage

kamp_variance(
  ppp_obj,
  rvec = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  markvar = "immune",
  thin_pct = 0
)

Arguments

ppp_obj

A point pattern object of class "ppp" from the spatstat package.

rvec

A vector of radii at which to calculate the KAMP variance.

correction

Type of edge correction. Defaults to translational.

markvar

The variable used to mark the points in the point pattern object. Default is "immune".

thin_pct

Percentage that determines how much to thin the amount of points in the point pattern object.

Value

A dataframe with the following columns:

r

The radius at which K was calculated.

k

The observed K value

theo_csr

The theoretical K under CSR

kamp_csr

The adjusted CSR representing the KAMP permuted expectation.

kamp_fundiff

The difference between observed K and KAMP CSR

var

Variance of K under the permutation null distribution

z

Z statistic, calculated by normalizing K using the formula: (K - KAMP)/sqrt(var)

pval

P-value, calculated using the formula: pnorm(-z)

Details

KAMP Variance

Examples

if (requireNamespace("spatstat.geom", quietly = TRUE)) {
  # simulates a simple spatial point pattern with two types
  win <- spatstat.geom::owin(c(0, 1), c(0, 1))
  pp <- spatstat.random::rpoispp(lambda = 100, win = win)
  marks <- sample(c("immune", "background"), pp$n, replace = TRUE)
  marked_pp <- spatstat.geom::ppp(pp$x, pp$y, window = win, marks = factor(marks))

  # computes KAMP variance
  kamp_result <- kamp_variance(marked_pp, markvar = "immune")
  print(kamp_result)
}
#> # A tibble: 6 × 7
#>       r       k theo_csr kamp_csr        var       z   pvalue
#>   <dbl>   <dbl>    <dbl>    <dbl>      <dbl>   <dbl>    <dbl>
#> 1 0     0        0        0       0          NaN     NaN     
#> 2 0.05  0.00950  0.00785  0.00764 0.00000278   1.11    0.133 
#> 3 0.075 0.0229   0.0177   0.0186  0.00000699   1.63    0.0521
#> 4 0.1   0.0367   0.0314   0.0309  0.0000152    1.49    0.0683
#> 5 0.15  0.0752   0.0707   0.0673  0.0000370    1.30    0.0969
#> 6 0.2   0.129    0.126    0.122   0.0000639    0.835   0.202